docs: waitsendpay timestamps are u32
waitsendpay reports the payment via json_add_payment_fields(),
which emits created_at and completed_at with json_add_u32
(lightningd/pay.c:158,160; wallet/wallet.h:363-364).  completed_at
was typed number.

Changelog-None

+++ cln-rpc/src/model.rs
@@ -7482,7 +7482,7 @@ pub mod responses {
 	    #[serde(skip_serializing_if = "Option::is_none")]
 	    pub bolt12: Option<String>,
 	    #[serde(skip_serializing_if = "Option::is_none")]
-	    pub completed_at: Option<f64>,
+	    pub completed_at: Option<u32>,
 	    #[serde(skip_serializing_if = "Option::is_none")]
 	    pub destination: Option<PublicKey>,
 	    #[serde(skip_serializing_if = "Option::is_none")]
@@ -7496,7 +7496,7 @@ pub mod responses {
 	    // Path `WaitSendPay.status`
 	    pub status: WaitsendpayStatus,
 	    pub amount_sent_msat: Amount,
-	    pub created_at: u64,
+	    pub created_at: u32,
 	    pub created_index: u64,
 	    pub groupid: u64,
 	    pub id: u64,
